clean(sequences ec): moved EC number processing code to separate file
