+++ app/assets/javascripts/sequence_show.js
@@ -1,7 +1,7 @@
 import {addCopy, logToGoogle, triggerDownloadModal, stringTitleize} from "./utils.js";
 import {showInfoModal} from "./modal.js";
 import {AmountTable} from "./components/amounttable.js";
-import PriorityQueue from "./utilities/PriorityQueue.js";
+import ECNumbers from "./components/ec_numbers.js";
 import "unipept-visualizations/src/treeview/treeview.js";
 
 /* eslint require-jsdoc: off */
@@ -148,115 +148,8 @@ function initSequenceShow(data) {
      * @return {TreeView} The created treeview
      */
     function setUPECTree({numAnnotatedPeptides, data}) {
-        /* Function to create a compareable string from EC numbers*/
-        let makeId = code =>{
-            // Pad each part with zeros so we can sort alphabeticly
-            // (there are EC number with letters)
-            const values = code.split(".").map(x => ("0000"+x).slice(-4));
-            return values.join(".");
-        };
-
-        // A map to store pointers to places in the tree that have already been
-        // created (non changing so we use object)
-        let map = Object.create(null);
-
-        // We make a root node, and add it to the map
-        let results = {id: 0,
-            name: "-",
-            fullname: "Enzyme Commission numbers",
-            children: [],
-            data: {self_count: 0},
-        };
-
-        map["-"] = results;
-
-        // Sort from general to specific
-        const sortedEC = Array.from(data.values()).sort((a, b) => (a.code+".-").split(".").indexOf("-") - (b.code+".-").split(".").indexOf("-"));
-
-        for (let data of sortedEC) {
-            let {code, value: count, name} = data;
-            const tmpPath = code.split(".").filter(x => x!=="-");
-            const shortCode = tmpPath.join("."); // "-" at end are now removed
-
-            let path = new Array(tmpPath.length);
-            // "3.1.15.24" => ["3.1.15", "3.1", "3", "-"]
-            for (let i = tmpPath.length; i>0; i--) {
-                path[tmpPath.length-i] = tmpPath.slice(0, i-1).join(".");
-            }
-            path[tmpPath.length-1] = "-";
-
-            // Create a node for the new EC-code and place it in the map
-            let placed = false;
-            let toInsert = {
-                id: makeId(shortCode),
-                name: shortCode,
-                fullname: name,
-                children: [],
-                data: {self_count: count, count: count, data: data},
-            };
-
-            map[shortCode] = toInsert;
-            // visit ancestors: ["3.1.15", "3.1", "3", "-"]
-            // and iteratively add `toInsert` as a child of its parent
-            for (let c of path) {
-                if (c in map) {
-                    if (!placed) map[c].children.push(toInsert);
-                    map[c].data.count += toInsert.data.count;
-                    placed = true;
-                } else {
-                    map[c] = {
-                        id: makeId(c),
-                        name: c,
-                        children: [toInsert],
-                        data: {self_count: 0, count: toInsert.data.count},
-                    };
-                    toInsert = map[c];
-                }
-            }
-        }
-
-        // Order the nodes by their id (order by EC number)
-        Object.values(map).forEach(obj => obj.children.sort((a, b) => a.id.localeCompare(b.id)));
-
-        // Finally create tree
-        let tree= $("#ec-treeview").empty().treeview(results, {
-            width: 916,
-            height: 600,
-            levelsToExpand: 1,
-            getTooltip: d => {
-                let fullcode = "EC:"+ (d.name + ".-.-.-.-").split(".").splice(0, 4).join(".");
-                let tip = `<strong>${fullcode}</strong>`;
-                if ("data" in d.data) {
-                    tip += `<br>${d.data.data.name}`;
-                }
-                tip += `<br>ocurrences: ${d.data.count} (${(100*d.data.count/numAnnotatedPeptides).toFixed(1)}%) <br>`;
-                if (d.data.count == d.data.self_count) {
-                    tip += "All specific";
-                } else {
-                    tip += `Specific ocurrences: ${d.data.self_count} (${(100*d.data.self_count/numAnnotatedPeptides).toFixed(1)}%)`;
-                }
-                return tip;
-            },
-        });
-
-        // expand certain nodes
-        // iteratively open the leaf with the largest count
-        // TODO: move to somewhere else
-        let root = tree.getRoot();
-        let allowedCount = root.data.count*2;
-        let pq = new PriorityQueue((a, b) => b.data.count - a.data.count);
-        root.children.forEach(c => pq.add(c));
-        while (allowedCount > 0) {
-            let toExpand = pq.remove();
-            allowedCount -= toExpand.data.count;
-            toExpand.expand(1);
-            (toExpand.children || []).forEach(c => pq.add(c));
-        }
-        tree.update(root);
-
-        // HACK: place the tree more to the left so everything is visible
-        setTimeout(()=>d3.select("#ec-treeview>svg>g>g").attr("transform", "translate(85,295)"), 1000);
-
+        let ecOrganiser = new ECNumbers(data);
+        let tree = ecOrganiser.createTree();
         // save tree
         $("#save-btn-ec").click(() => {
             logToGoogle("Single Peptide", "Save EC Image");
